Refactor/refactor-user-test (#1)
* feat: 🎸 add OpenTelemetry dependencies for enhanced observability

- Introduced new dependencies for OpenTelemetry, including @elysiajs/opentelemetry, @opentelemetry/sdk-trace-node, and related packages to support tracing and logging.
- Updated package.json and pnpm-lock.yaml to reflect the addition of these dependencies, ensuring improved observability in the application.

* feat: 🎸 update base URL for development environment and add user API documentation

- Changed the base URL in the development environment settings from http://localhost:3000 to http://localhost:7000 for consistency.
- Introduced a new user.http file containing API documentation for user-related operations, including creating, retrieving, and updating users.

* feat: 🎸 refactor user DTO imports and create new user DTO file

- Updated import paths for user-related DTOs in CreateUserController, GetAllUsersController, and GetUserByIdController to point to the new dtos/user.dto.ts file.
- Created a new user.dto.ts file to define and export user-related DTOs, including CreateUserRequestDto, CreateUserResponseDto, GetUserResponseDto, and GetUsersResponseDto, enhancing code organization and maintainability.

* feat: 🎸 implement get user by ID use case and corresponding tests

- Added getUserByIdUseCase to handle user retrieval logic.
- Implemented error handling for non-existent users with NotFoundError.
- Created unit tests for getUserByIdUseCase to validate functionality and error scenarios.
- Refactored delete-user-by-id test to use promise for better readability.
