+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
   "rest-client.environmentVariables": {
     "development": {
-      "baseUrl": "http://localhost:3000",
+      "baseUrl": "http://localhost:7000",
       "contentType": "application/json"
     },
     "production": {
