+++ src/adapters/users/get-all-users.controller.ts
@@ -1,10 +1,10 @@
 import Elysia from 'elysia';
 import { inject, injectable } from 'tsyringe';
 import { GetAllUsersUseCase } from '../../core/domain/users/use-case/get-all-users.usecase';
-import { GetUsersResponseDto } from '../../core/shared/dtos/user.dto';
 import { UserMapper } from './mappers/user.mapper';
 import { TOKENS } from '../../core/shared/tokens';
 import type { LoggerPort } from '../../core/shared/logger/logger.port';
+import { GetUsersResponseDto } from './dtos/user.dto';
 
 @injectable()
 export class GetAllUsersController {
