+++ src/core/domain/users/use-case/get-user-by-id.usecase.spec.ts
@@ -1,18 +1,18 @@
 import 'reflect-metadata';
+import { UserRepository } from '../service/user.repository';
+import { getUserByIdUseCase } from './get-user-by-id.usecase';
 import { faker } from '@faker-js/faker';
 import { vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 import { IUser, UserId } from '../entity/user.entity';
-import { UserRepository } from '../service/user.repository';
 import { NotFoundError } from 'elysia';
-import { DeleteUserByIdUseCase } from './delete-user-by-id.usecase';
 
-describe('DeleteUserByIdUseCase', () => {
-  let useCase: DeleteUserByIdUseCase;
+describe('getUserByIdUseCase', () => {
+  let useCase: getUserByIdUseCase;
   const userRepository = mock<UserRepository>();
 
   beforeEach(() => {
-    useCase = new DeleteUserByIdUseCase(userRepository);
+    useCase = new getUserByIdUseCase(userRepository);
   });
 
   afterEach(() => {
@@ -34,17 +34,19 @@ describe('DeleteUserByIdUseCase', () => {
     expect(userRepository.deleteById).not.toHaveBeenCalled();
   });
 
-  it('should be delete user', async () => {
+  it('should be get user by id', async () => {
     //Arrange
     const user = mock<IUser>({ id: userId });
     userRepository.findById.mockResolvedValue(user);
-    userRepository.deleteById.mockResolvedValue(true);
+
+    const expected = user;
 
     //Act
     const actual = await useCase.execute(userId);
+
     //Assert
-    expect(actual).toBeTruthy();
+    expect(actual).toEqual(expected);
     expect(userRepository.findById).toHaveBeenCalledWith(userId);
-    expect(userRepository.deleteById).toHaveBeenCalledWith(userId);
+    expect(userRepository.deleteById).not.toHaveBeenCalled();
   });
 });
