+++ src/core/domain/users/use-case/get-user-by-id.usecase.ts
@@ -4,7 +4,6 @@ import { IUser, UserId } from '../entity/user.entity';
 import { UserRepository } from '../service/user.repository';
 import { TOKENS } from '../../../shared/tokens';
 import { NotFoundError } from 'elysia';
-import { ReasonPhrases } from 'http-status-codes';
 
 @injectable()
 export class getUserByIdUseCase implements IUseCase<UserId, IUser> {
@@ -15,7 +14,7 @@ export class getUserByIdUseCase implements IUseCase<UserId, IUser> {
 
   async execute(userId: UserId): Promise<IUser> {
     const user = await this.userRepository.findById(userId);
-    if (!user) throw new NotFoundError(ReasonPhrases.NOT_FOUND);
+    if (!user) throw new NotFoundError('User not found');
     return user;
   }
 }
